refactor: move the packages under internal and the command under cmd
Nothing here was ever meant to be imported from another module, and internal/
is how the compiler says so rather than a paragraph asking politely. Both
packages are now free to change shape without it being someone else's problem.

The command moves in the same step, because a main package sitting at the root
is what made them look importable in the first place. The directory name is the
binary name under `go install`, so cmd/gwttr rather than cmd/cli keeps that
producing something called gwttr.

Two things downstream break, both documented in the README: `go install
github.com/alrayyes/gwttr@latest` needs the /cmd/gwttr suffix now, and neither
package can be imported at all. Nothing is known to be doing either.

goreleaser needs telling where main is; the archives it builds are unchanged.
go.mod picks up a correction on the way past, since cobra was marked indirect
while main.go imported it directly.
