+++ internal/weather/weather_test.go
@@ -8,8 +8,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/alrayyes/gwttr/weather"
-	"github.com/alrayyes/gwttr/wttrclient"
+	"github.com/alrayyes/gwttr/internal/weather"
+	"github.com/alrayyes/gwttr/internal/wttrclient"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
