+++ internal/wttrclient/wttrclient_test.go
@@ -5,7 +5,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/alrayyes/gwttr/wttrclient"
+	"github.com/alrayyes/gwttr/internal/wttrclient"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
